pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aba_optimiser"
version = "0.1.0"
description = "Knob strength optimisation using PyMAD-NG"
authors = [
    { name="Your Name", email="you@example.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "pandas",
    "tfs-pandas",
    "tensorboardX",
    "pymadng>=0.3.0",
    "matplotlib",
    "uncertainties",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov"
]
docs = [
    "sphinx>=7.0",
    "sphinx-autodoc-typehints>=2.0",
    "myst-parser>=2.0",
    "furo>=2023.9.10",
]
tracking = [
    "xsuite",
    "omc3 > 0.26.0",
    "cpymad",
    "pyarrow",
    "psutil",
]
cern = [
    "nxcals",
]
all = [
    "test",
    "dev",
    "docs",
    "tracking",
    "cern",
]


[tool.setuptools]
packages = ["aba_optimiser"]
package-dir = {"" = "src"}

[tool.coverage.run]
source = ["src/aba_optimiser"]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["tests/data"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
addopts = [
    "--log-cli-level=INFO",
    "--log-level=INFO",
    "--log-cli-format=%(levelname)s %(name)s %(message)s",
    "-s",
]


# ----- Dev Tools Configuration ----- #

[tool.ruff]
exclude = [
  ".eggs",
  ".git",
  ".mypy_cache",
  ".venv",
  "_build",
  "build",
  "dist",
]

# Assume Python 3.10+
target-version = "py310"

line-length = 100
indent-width = 4

[tool.ruff.lint]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
ignore = [
  "E501",  # line too long
  "FBT001",  # boolean-type-hint-positional-argument
  "FBT002",  # boolean-default-value-positional-argument
  "PT019",  # pytest-fixture-param-without-value (but suggested solution fails)
]
extend-select = [
    "F",        # Pyflakes rules
    "W",        # PyCodeStyle warnings
    "E",        # PyCodeStyle errors
    "I",        # Sort imports properly
    "A",        # Detect shadowed builtins
    "N",        # enforce naming conventions, e.g. ClassName vs function_name
    "UP",       # Warn if certain things can change due to newer Python versions
    "C4",       # Catch incorrect use of comprehensions, dict, list, etc
    "FA",       # Enforce from __future__ import annotations
    "FBT",      # detect boolean traps
    "ISC",      # Good use of string concatenation
    "BLE",      # disallow catch-all exceptions
    "ICN",      # Use common import conventions
    "RET",      # Good return practices
    "SIM",      # Common simplification rules
    "TID",      # Some good import practices
    "TC",       # Enforce importing certain types in a TYPE_CHECKING block
    "PTH",      # Use pathlib instead of os.path
    "NPY",      # Some numpy-specific things
]
# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []


[tool.ty.src]
exclude = [
    "**/acc-models-lhc/",
    "*_analysis*/",
    "*_results/",
    "analysis*/",
    "b1_flat_60_18cm/",
    "betabeat/",
    "data/",
    "docs/api/",
    "/mad_scripts/",
    "models/**",
    "plots/",
    "*.out",
    "*.png",

]